+++ CHANGES.rst
@@ -9,6 +9,8 @@ Changelog of lizard-portal
 
 - bugfix for organization name in user information (pp issue 139)
 
+- Improved analyse-interpretatie-grid
+
 
 0.38 (2012-04-12)
 -----------------